+++ src/pages/home/subPages/hooks/index.ts
@@ -1,2 +1,3 @@
 export { default as usePressEnterFetch } from './usePressEnterFetch';
 export { default as useIntersectionObserver } from './useIntersectionObserver';
+export { useHorizontalScroll } from './useHorizontalScroll';
